refactor/change fetch to axios
