+++ package.json
@@ -11,6 +11,7 @@
   "dependencies": {
     "@hookform/resolvers": "^3.1.1",
     "@radix-ui/react-accordion": "^1.1.2",
+    "axios": "^1.4.0",
     "clsx": "^1.2.1",
     "next": "13.4.8",
     "react": "18.2.0",
