+++ src/components/contact/contactForm/useContactForm.ts
@@ -1,4 +1,4 @@
-import { mutateData } from '@/utils/mutateData';
+import { mutateData } from '@/utils/axios';
 import type { ContactFormValues } from '@/utils/schemas/contactFormSchema';
 import { useMutation } from 'react-query';
 
@@ -6,12 +6,7 @@ type ContactFormPayload = Pick<ContactFormValues, 'email' | 'name' | 'message'>;
 
 export const useContactForm = () => {
 	const { data, isLoading, isSuccess, isError, error, mutate } = useMutation(
-		(payload: ContactFormPayload) =>
-			mutateData({
-				path: 'https://uki.edu.pl/api/contact',
-				method: 'POST',
-				body: payload,
-			})
+		(payload: ContactFormPayload) => mutateData('/api/contact', payload)
 	);
 
 	const submit = ({
