--- src/utils/fetchData.ts
@@ -1,9 +1,0 @@
-export const fetchData = async (url: string) => {
-	const response = await fetch(url);
-
-	if (!response.ok) {
-		throw new Error('Failed to fetch data');
-	}
-
-	return await response.json();
-};
